refactor(utils): import submodules directly to avoid heavy __init__ overhead
