+++ src/novel_downloader/__main__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+
+from novel_downloader.cli import cli_main
+
+if __name__ == "__main__":
+    cli_main()
