+++ src/novel_downloader/core/downloaders/common.py
@@ -13,7 +13,8 @@ from typing import Any
 from novel_downloader.core.downloaders.base import BaseDownloader
 from novel_downloader.core.downloaders.signals import STOP, Progress, StopToken
 from novel_downloader.models import BookConfig, ChapterDict
-from novel_downloader.utils import ChapterStorage, async_jitter_sleep
+from novel_downloader.utils.chapter_storage import ChapterStorage
+from novel_downloader.utils.time_utils import async_jitter_sleep
 
 
 class CommonDownloader(BaseDownloader):
