+++ src/novel_downloader/core/exporters/base.py
@@ -16,7 +16,8 @@ from pathlib import Path
 from typing import Any, ClassVar, Self, cast
 
 from novel_downloader.models import BookInfoDict, ChapterDict, ExporterConfig
-from novel_downloader.utils import ChapterStorage, get_cleaner
+from novel_downloader.utils.chapter_storage import ChapterStorage
+from novel_downloader.utils.text_utils import get_cleaner
 
 
 class SafeDict(dict[str, Any]):
