+++ src/novel_downloader/core/parsers/qidian.py
@@ -25,14 +25,11 @@ from novel_downloader.models import (
     ParserConfig,
     VolumeInfoDict,
 )
-from novel_downloader.utils import (
-    download,
-    truncate_half_lines,
-)
 from novel_downloader.utils.constants import DATA_DIR
 from novel_downloader.utils.cookies import get_cookie_value
 from novel_downloader.utils.fontocr import get_font_ocr
 from novel_downloader.utils.node_decryptor import get_decryptor
+from novel_downloader.utils.text_utils import truncate_half_lines
 
 logger = logging.getLogger(__name__)
 
@@ -263,6 +260,8 @@ class QidianParser(BaseParser):
                 cid,
             )
             return ""
+
+        from novel_downloader.utils.network import download
 
         css_str = chapter_info.get("css")
         random_font_str = chapter_info.get("randomFont")
